requirements.txt
aiohttp==3.6.2
alembic==1.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.3
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonlib-python3==1.6.1
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
mysqlclient==1.4.6
pkg-resources==0.0.0
pyee==7.0.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
slackclient==2.7.1
slackeventsapi==2.1.0
SQLAlchemy==1.3.17
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
yarl==1.4.2